IOS signiture bug fixed.

+++ Assets/Scripts/ScreenMarkerIOSPlugin.cs
@@ -21,14 +21,14 @@ public class ScreenMarkerIOSPlugin : MonoBehaviour, IScreenMarker
         int x, int y, int width, int height,
         string text, string fontName, float fontSize,
         string colorString,
-        float angle,
+        int angle,
         bool useSizeToFit);
     [DllImport("__Internal")]
     private static extern void _AddTextWithCenter(
         int x, int y,
         string text, string fontName, float fontSize,
         string colorString,
-        float angle);
+        int angle);
     [DllImport("__Internal")]
     private static extern void _ClearTextAll();
     [DllImport("__Internal")]
@@ -45,7 +45,7 @@ public class ScreenMarkerIOSPlugin : MonoBehaviour, IScreenMarker
         string fontName,
         float fontSize,
         string colorString,
-        float angle,
+        int angle,
         int horizontalMargin, int verticalMargin);
     [DllImport("__Internal")]
     private static extern void _UnsetTextTileMode();
@@ -60,7 +60,7 @@ public class ScreenMarkerIOSPlugin : MonoBehaviour, IScreenMarker
     private static extern void _SetImageTileMode(
         byte[] imageBytes,
         int length,
-        float angle,
+        int angle,
         int horizontalMargin, int verticalMargin);
     [DllImport("__Internal")]
     private static extern void _SetImageTileModeWithText(
@@ -69,7 +69,7 @@ public class ScreenMarkerIOSPlugin : MonoBehaviour, IScreenMarker
         string fontName,
         float fontSize,
         string colorString,
-        float angle,
+        int angle,
         int horizontalMargin, int verticalMargin);
     [DllImport("__Internal")]
     private static extern void _UnsetImageTileMode();
